AwsGameKit/Resources/cloudResources/requirements.txt
# AWS Lambda - Python 3.7 Runtime: https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html
# Use the latest version, because the AWS Lambda runtimes are upgraded over time.
boto3
botocore

# Crypto Lambda Layer:
cryptography == 3.4.7
python-jose[cryptography] == 3.3.0

# Image Processing Lambda Layer:
Pillow == 9.0.1
